fix builder & warning

+++ BTL/clientserver.cc
@@ -325,7 +325,7 @@ void Sockpeer::run(){
                 struct sockaddr_in peerAddress;
 
                 n = this->networkObj->networkRecv(charData, this->BUFFSIZE, &peerAddress);
-                if (n == -1){
+                if (n == (size_t)-1){
                     // Something wrong with the data, ask to send again?
                     continue;
                 }
@@ -566,7 +566,7 @@ void Sockpeer::run(){
                     if (fileHandle == 0){
                         continue;
                     }
-                    if (reply.status() == -1){
+                    if (reply.status() == (size_t)-1){
                         // Some peer is done, update it to seeder
                         // printf("Getting %s:%d\n", peerHost.c_str(), peerPort);
                         bool isDone = true;
